+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+##
+
+* change `plot_eval_results` to show individual metrics' results as subplots
+
 ## 0.4.2 - 2018-02-06
 
 * made greedy partitioning much more efficient (i.e. faster work distribution)